Refactor impl_set_plan.rs into impl_coordinator_channel.rs
