+++ src/work_unit_feed/mod.rs
@@ -5,7 +5,7 @@ mod work_unit_feed;
 mod work_unit_feed_provider;
 mod work_unit_feed_registry;
 
-pub(crate) use remote_work_unit_feed::{RemoteWorkUnitFeedRegistry, RemoteWorkUnitFeedTxs};
+pub(crate) use remote_work_unit_feed::RemoteWorkUnitFeedRegistry;
 pub(crate) use work_unit_feed_registry::{WorkUnitFeedRegistry, set_distributed_work_unit_feed};
 
 pub use work_unit::WorkUnit;
