+++ src/worker/mod.rs
@@ -1,9 +1,10 @@
 pub(crate) mod generated;
+mod impl_coordinator_channel;
 mod impl_execute_task;
-mod impl_set_plan;
 mod session_builder;
 mod single_write_multi_read;
 mod spawn_select_all;
+mod task_data;
 #[cfg(any(test, feature = "integration"))]
 pub(crate) mod test_utils;
 mod worker_connection_pool;
@@ -18,4 +19,4 @@ pub use session_builder::{
 };
 pub use worker_service::Worker;
 
-pub use impl_set_plan::TaskData;
+pub use task_data::TaskData;
